+++ resources/JenkinsfileTemplate.groovy
@@ -171,7 +171,6 @@ pipeline {
             name 'PLATFORM'
             values (
               'debian-9',
-              'ubuntu-16',
               'ubuntu-18',
               'ubuntu-20'
               )
@@ -183,7 +182,7 @@ pipeline {
           */
           stage('Build') {
             steps {
-              buildUnix()
+              runBuild()
             }
           }
           /**
@@ -241,18 +240,19 @@ pipeline {
           axis {
             name 'PLATFORM'
             values (
-              'windows-2016-immutable',
-              'windows-2016-latest-immutable',
               'windows-2012-r2-immutable',
-              'windows-2012-r2-latest-immutable',
+              'windows-2016-immutable',
               'windows-2019-immutable',
-              'windows-2019-docker-immutable',
-              'windows-2019-test-immutable',
-              'windows-2019-latest-immutable'
+              'windows-2019-docker-immutable'
               )
           }
         }
         stages {
+          stage('Build') {
+            steps {
+              runBuild()
+            }
+          }
           stage('Test') {
             steps {
               testWindows()
@@ -291,7 +291,7 @@ pipeline {
         stage('Mac OS X check - 01'){
           agent { label 'worker-c07yx0vrjyvy' }
           steps {
-            buildUnix()
+            runBuild()
             testMac()
           }
           post {
@@ -303,7 +303,7 @@ pipeline {
         stage('Mac OS X check - 02'){
           agent { label 'worker-c07yx0vdjyvy' }
           steps {
-            buildUnix()
+            runBuild()
             testMac()
           }
           post {
@@ -315,7 +315,7 @@ pipeline {
         stage('BareMetal worker-854309 check'){
           agent { label 'worker-854309' }
           steps {
-            buildUnix()
+            runBuild()
             testBaremetal()
           }
           post {
@@ -327,7 +327,7 @@ pipeline {
         stage('BareMetal worker-1095690 check'){
           agent { label 'worker-1095690' }
           steps {
-            buildUnix()
+            runBuild()
             testBaremetal()
           }
           post {
@@ -339,7 +339,7 @@ pipeline {
         stage('BareMetal worker-1213919 check'){
           agent { label 'worker-1213919' }
           steps {
-            buildUnix()
+            runBuild()
             testBaremetal()
           }
           post {
@@ -351,7 +351,7 @@ pipeline {
         stage('BareMetal worker-1225339 check'){
           agent { label 'worker-1225339' }
           steps {
-            buildUnix()
+            runBuild()
             testBaremetal()
           }
           post {
@@ -400,11 +400,15 @@ def testDockerInside(){
   }
 }
 
-def buildUnix(){
+def runBuild(){
   deleteDir()
   unstash 'source'
   dir("${BASE_DIR}"){
-    sh returnStatus: true, script: './resources/scripts/jenkins/build.sh'
+    if (isUnix()) {
+      sh(returnStatus: true, script: './resources/scripts/jenkins/build.sh')
+    } else {
+      bat(returnStatus: true, script: '.\\resources\\scripts\\jenkins\\build.bat')
+    }
   }
 }
 
@@ -440,16 +444,6 @@ def testWindows(params = [:]){
     dir("${BASE_DIR}"){
       powershell(script: ".\\resources\\scripts\\jenkins\\apm-ci\\test.ps1 ${withExtra}")
     }
-  } else {
-    checkOldWindows()
-  }
-}
-
-def checkOldWindows(){
-  deleteDir()
-  unstash 'source'
-  dir("${BASE_DIR}"){
-    bat(returnStatus: true, script: '.\\resources\\scripts\\jenkins\\build.bat')
   }
 }